Fix issue when adding hunt prey options
The existing options in the roll params can be either an array or a set,
but the existing code, which uses `.append()`, doesn't work on a set.

Adding to the passed in options array/set also leaves it modified after the
function rerturns.  If the caller re-uses it, for instance rolling for the
whole party, it will have had options like hunted-prey added to it on
subsequent rolls.  The pf2e system code doesn't modify the parameters.

Convert an existing option list to a new set and add the new options to it,
which is what the pf2e system code does, to solve both problems at once.